Internal version checking.
Each library reports it's version.  External code bases can use this ability to 
ensure that they have expected code, and can die if it doesn't match, to avoid 
any assumptions about functionality.  It adds the inherent possibility of being 
incorrect or out-of-date, but I'm hoping it will be worth the effort in the end.

The reason it has been added is for cs-project (which I'm working to get to the 
point of uploading to sourceforge.net): it requires functionality in 
cs_fileSystemClass that isn't there in previous versions, and the system 
simply wouldn't work without it...

The following class files now have a public "get_version()" method:
	* contentSystemClass.php
	* cs_fileSystemClass.php
	* cs_genericPageClass.php
	* cs_globalFunctions.php
	* cs_sessionClass.php
	* cs_tabsClass.php
/contentSystemClass.php:
	* initialize_locals():
		-- checks that cs_genericPage, cs_fileSystemClass, and cs_globalFunctions 
		all have the same version to avoid problems (and to ensure they're kept 
		up-to-date). :)