Re-add cs_globalFunctions::get_version().  AGAIN.  FOR REAL.
It appears that I went CRAZY when I reversed part of the patch that reversed... 
oh, I don't know what the bloody hell I did.  Committed a version that had 
get_version() but lacked a bunch of fixes, then re-added all the fixes and 
removed get_version()... It's back.  If I have to resurrect it again, it just 
goes away for good. :|